7.2.5修正 intで出力

+++ Chapter07/演習問題/Exercise02/Program.cs
@@ -59,7 +59,7 @@ namespace Exercise02 {
 
         private static void Exercise5(List<Book> books) {
             var a = books.Where(n => n.Price < 4000);
-            Console.WriteLine(a.Max(n => n.Pages));
+            Console.WriteLine((int)a.Max(n => n.Pages));
         }
 
         private static void Exercise6(List<Book> books) {
